+++ desktop-native/App.xaml.cs
@@ -1,0 +1,3 @@
+using System.Windows;
+namespace MiniChat.Native;
+public partial class App : Application { }